Fix: eliminate -Wimplicit-int-conversion, -Wfloat-conversion, and -Wcast-align warnings (Cluster J)
- dkim-dns.c: replace (HEADER *) buffer overlay with memcpy/local-struct to fix
  cast-align (alignment 1→4 on non-x86 architectures)
- config.c: fgetc() int→char narrowing — explicit (char) cast
- test.c: time_t % 65536 → uint16_t truncation — explicit (uint16_t) cast
- t-test136.c: tolower() int→char — (char) tolower((unsigned char) *p)
- miltertest.c: 12 × lua_tonumber() → int/unsigned-long/uint32_t implicit
  float→int conversions — explicit casts to declared target types
- miltertest.c: servent.s_port int→in_port_t and strtoul int→uint16_t port
  assignments — explicit casts
- miltertest.c: sockaddr*→sockaddr_in/in6* cast-align — route through void*
  (getaddrinfo guarantees alignment; void* intermediate suppresses the check)

+++ opendkim/test.c
@@ -891,7 +891,7 @@ dkimf_testfiles(DKIM_LIB *libopendkim, char *flist, uint64_t fixedtime,
 
 	(void) memset(&sin, '\0', sizeof sin);
 	sin.sin_family = AF_INET;
-	sin.sin_port = htons(time(NULL) % 65536);
+	sin.sin_port = htons((uint16_t)(time(NULL) % 65536));
 	sin.sin_addr.s_addr = htonl(INADDR_LOOPBACK);
 
 	/*
